chore: reorganizing the screens into inline components and removing unused code
